+++ ogre2/src/media/materials/programs/gpu_rays_1st_pass_fs.glsl
@@ -25,6 +25,8 @@ in block
 
 uniform sampler2D depthTexture;
 uniform sampler2D colorTexture;
+uniform sampler2D particleDepthTexture;
+uniform sampler2D particleTexture;
 
 out vec4 fragColor;
 
@@ -34,19 +36,45 @@ uniform float far;
 uniform float min;
 uniform float max;
 
-float getDepth(vec2 uv)
+uniform float particleStddev;
+uniform float particleScatterRatio;
+uniform float time;
+
+// see gaussian_noise_fs.glsl for documentation on the rand and gaussrand
+// functions
+
+#define PI 3.14159265358979323846264
+
+float rand(vec2 co)
 {
-  float fDepth = texture(depthTexture, uv).x;
-  float linearDepth = projectionParams.y / (fDepth - projectionParams.x);
-  return linearDepth;
+  float r = fract(sin(dot(co.xy, vec2(12.9898,78.233))) * 43758.5453);
+  // Make sure that we don't return 0.0
+  if(r == 0.0)
+    return 0.000000000001;
+  else
+    return r;
 }
 
+vec4 gaussrand(vec2 co, vec3 offsets, float stddev, float mean)
+{
+  float U, V, R, Z;
+  U = rand(co + vec2(offsets.x, offsets.x));
+  V = rand(co + vec2(offsets.y, offsets.y));
+  R = rand(co + vec2(offsets.z, offsets.z));
+  if(R < 0.5)
+    Z = sqrt(-2.0 * log(U)) * sin(2.0 * PI * V);
+  else
+    Z = sqrt(-2.0 * log(U)) * cos(2.0 * PI * V);
+  Z = Z * stddev + mean;
+  return vec4(Z, Z, Z, 0.0);
+}
 
 void main()
 {
   // get linear depth
-  float d = getDepth(inPs.uv0);
-  
+  float fDepth = texture(depthTexture, inPs.uv0).x;
+  float d = projectionParams.y / (fDepth - projectionParams.x);
+
   // get retro
   float retro = texture(colorTexture, inPs.uv0).x * 2000.0;
 
@@ -55,6 +83,27 @@ void main()
 
   // get length of 3d point, i.e.range
   float l = length(viewSpacePos);
+
+  // particle mask - color and depth
+  vec4 particle = texture(particleTexture, inPs.uv0);
+  float particleDepth = texture(particleDepthTexture, inPs.uv0).x;
+
+  if (particle.x > 0.0 && particleDepth > 0.0 && particleDepth < fDepth)
+  {
+    // apply scatter effect so that only some of the smoke pixels are visible
+    float r = rand(inPs.uv0 + vec2(time, time));
+    if (r < particleScatterRatio)
+    {
+      float pd = projectionParams.y / (particleDepth - projectionParams.x);
+      vec3 point = inPs.cameraDir * pd;
+
+      // apply gaussian noise to particle depth data
+      point = point + gaussrand(inPs.uv0, vec3(time, time, time),
+          particleStddev, 0.0).xyz;
+
+      l = length(point);
+    }
+  }
 
   if (l > far)
     l = max;